# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
export type HealthStatusStatus = typeof HealthStatusStatus[keyof typeof HealthStatusStatus];


export const HealthStatusStatus = {
  ok: 'ok',
  degraded: 'degraded',
} as const;

export interface HealthStatus {
  status: HealthStatusStatus;
}

export interface SystemStatus {
  name: string;
  version: string;
  uptimeSeconds: number;
  memoryUsageMb: number;
  nodeVersion: string;
  environment: string;
}

export interface Model {
  id: string;
  name: string;
  description: string;
  provider: string;
  contextWindow: number;
  maxOutputTokens: number;
  capabilities: string[];
  createdAt: string;
}

export interface ModelList {
  object: string;
  data: Model[];
  total: number;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  system: 'system',
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatCompletionRequest {
  model: string;
  /** @minItems 1 */
  messages: ChatMessage[];
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
  /**
     * @minimum 1
     * @maximum 32768
     */
  maxTokens?: number;
  stream?: boolean;
}

export type ChatCompletionChoiceFinishReason = typeof ChatCompletionChoiceFinishReason[keyof typeof ChatCompletionChoiceFinishReason];


export const ChatCompletionChoiceFinishReason = {
  stop: 'stop',
  length: 'length',
  content_filter: 'content_filter',
} as const;

export interface ChatCompletionChoice {
  index: number;
  message: ChatMessage;
  finishReason: ChatCompletionChoiceFinishReason;
}

export interface ChatUsage {
  promptTokens: number;
  completionTokens: number;
  totalTokens: number;
}

export interface ChatCompletionResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: ChatCompletionChoice[];
  usage: ChatUsage;
}

export type ErrorResponseErrorDetails = { [key: string]: unknown };

export type ErrorResponseError = {
  code: string;
  message: string;
  details?: ErrorResponseErrorDetails;
};

export interface ErrorResponse {
  error: ErrorResponseError;
}