# lib/api-zod/src/generated/types/chatCompletionChoice.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { ChatCompletionChoiceFinishReason } from './chatCompletionChoiceFinishReason';
import type { ChatMessage } from './chatMessage';

export interface ChatCompletionChoice {
  index: number;
  message: ChatMessage;
  finishReason: ChatCompletionChoiceFinishReason;
}