# lib/api-zod/src/generated/types/chatCompletionRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { ChatMessage } from './chatMessage';

export interface ChatCompletionRequest {
  model: string;
  /** @minItems 1 */
  messages: ChatMessage[];
  /**
     * @minimum 0
     * @maximum 2
     */
  temperature?: number;
  /**
     * @minimum 1
     * @maximum 32768
     */
  maxTokens?: number;
  stream?: boolean;
}