# lib/api-zod/src/generated/types/chatCompletionResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { ChatCompletionChoice } from './chatCompletionChoice';
import type { ChatUsage } from './chatUsage';

export interface ChatCompletionResponse {
  id: string;
  object: string;
  created: number;
  model: string;
  choices: ChatCompletionChoice[];
  usage: ChatUsage;
}