# lib/api-zod/src/generated/types/chatMessage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { ChatMessageRole } from './chatMessageRole';

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}