# lib/api-zod/src/generated/types/chatUsage.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */

export interface ChatUsage {
  promptTokens: number;
  completionTokens: number;
  totalTokens: number;
}