# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { ErrorResponseError } from './errorResponseError';

export interface ErrorResponse {
  error: ErrorResponseError;
}