# lib/api-zod/src/generated/types/modelList.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */
import type { Model } from './model';

export interface ModelList {
  object: string;
  data: Model[];
  total: number;
}