# lib/api-zod/src/generated/types/systemStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * QuantumAI-Backend API specification
 * OpenAPI spec version: 1.0.0
 */

export interface SystemStatus {
  name: string;
  version: string;
  uptimeSeconds: number;
  memoryUsageMb: number;
  nodeVersion: string;
  environment: string;
}